pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "opensanctions"
version = "0.1.0"
description = "CLI and tools for processing OpenSanctions JSON-LD datasets"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "requests",
    "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
